Fix issue #53: Small changes are made to fix minor bugs.
Fix gradlew: gradlew is not found for some reason. So the imported project is ran and tested with gradle.
Fix folderPath: The path was given hard-coded in "runTests". It is changed to the folderPath parameter.
Fix: A minor bug causing the result strings of tests and compilation to start with "null" is resolved.